Use the ES6
